refactor(bc): finalize changes
